pyproject.toml
[project]
name = "filesystem-frontmatter-mcp"
version = "0.1.0"
description = "MCP server for querying Markdown frontmatter with DuckDB SQL"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.0.0",
  "duckdb>=1.0.0",
  "python-frontmatter>=1.0.0",
  "pyarrow>=22.0.0",
]

[project.scripts]
filesystem-frontmatter-mcp = "filesystem_frontmatter_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/filesystem_frontmatter_mcp"]

[dependency-groups]
dev = ["ruff>=0.14.6", "pytest>=8.0.0"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
